fix(#MOSIP-27873): [Vijay] handle the PR comments and optimise
