+++ src/main/java/io/mosip/mimoto/dto/mimoto/BindingOtpRequestDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
 @Data
@@ -11,6 +12,8 @@ import javax.validation.constraints.NotNull;
 @NoArgsConstructor
 public class BindingOtpRequestDto {
     private String requestTime;
+
+    @Valid
     @NotNull
     private BindingOtpInnerReqDto request;
 }
