+++ src/main/resources/bootstrap.properties
@@ -1,7 +1,7 @@
 spring.cloud.config.uri=localhost
 spring.cloud.config.label=develop
 spring.profiles.active=local
-spring.cloud.config.name=mimoto,inji,application
+spring.cloud.config.name=mimoto,inji
 spring.application.name=mimoto
 
 management.endpoint.health.show-details=always
